requirements.txt
langchain
langchain-google-genai
langchain-community
pypdf
python-dotenv
